w3c-reports/12092014/gtk/runned-tests/dom/nodes/Document-createEvent-expected.txt


PASS CustomEvent should be an alias for CustomEvent. 
FAIL createEvent('CustomEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL customevent should be an alias for CustomEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('customevent') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL CUSTOMEVENT should be an alias for CustomEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('CUSTOMEVENT') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS Event should be an alias for Event. 
FAIL createEvent('Event') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL event should be an alias for Event. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('event') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL EVENT should be an alias for Event. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('EVENT') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS Events should be an alias for Event. 
FAIL createEvent('Events') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL events should be an alias for Event. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('events') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL EVENTS should be an alias for Event. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('EVENTS') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS HTMLEvents should be an alias for Event. 
FAIL createEvent('HTMLEvents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL htmlevents should be an alias for Event. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('htmlevents') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL HTMLEVENTS should be an alias for Event. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('HTMLEVENTS') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS MouseEvent should be an alias for MouseEvent. 
FAIL createEvent('MouseEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL mouseevent should be an alias for MouseEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('mouseevent') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL MOUSEEVENT should be an alias for MouseEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('MOUSEEVENT') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS MouseEvents should be an alias for MouseEvent. 
FAIL createEvent('MouseEvents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL mouseevents should be an alias for MouseEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('mouseevents') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL MOUSEEVENTS should be an alias for MouseEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('MOUSEEVENTS') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS UIEvent should be an alias for UIEvent. 
FAIL createEvent('UIEvent') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL uievent should be an alias for UIEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('uievent') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL UIEVENT should be an alias for UIEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('UIEVENT') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS UIEvents should be an alias for UIEvent. 
FAIL createEvent('UIEvents') should be initialized correctly. assert_equals: isTrusted should be initialized to false expected (boolean) false but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createEvent.html:31:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:45:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL uievents should be an alias for UIEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('uievents') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:46:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL UIEVENTS should be an alias for UIEvent. NotSupportedError: DOM Exception 9(stack: createEvent@[native code]
http://localhost:8800/dom/nodes/Document-createEvent.html:12:30
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:11:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
FAIL createEvent('UIEVENTS') should be initialized correctly. undefined is not an object (evaluating 'ev.type')(stack: http://localhost:8800/dom/nodes/Document-createEvent.html:17:21
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
testAlias@http://localhost:8800/dom/nodes/Document-createEvent.html:16:7
http://localhost:8800/dom/nodes/Document-createEvent.html:47:12
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Document-createEvent.html:44:10)
PASS Should throw NOT_SUPPORTED_ERR for unrecognized arguments 
